pub-sub-trigger-function/function-source-code/requirements.txt
# Function dependencies, for example:
# package>=version
pandas
google-cloud-storage
